+++ src/index.html
@@ -44,6 +44,7 @@
 
     <!-- Футер Міли -->
     <load src="./partials/footer.html" />
-    <script src="/src/main.js"></script>
+
+    <script type="module" src="./main.js"></script>
   </body>
 </html>
